Correcciones de las PR del primer y segundo Sprint
