+++ .gitignore
@@ -11,7 +11,6 @@ node_modules
 dist
 dist-ssr
 *.local
-yarn.lock
 
 # Editor directories and files
 .vscode/*
