--- integrantes.md
@@ -1,6 +1,0 @@
-- Paul Landaeta
-- Alexia Marin
-- Andres Sanchez
-- Miguel Quenta
-- Adrian Sanchez
-- Zein Tonconi
