+++ src/components/administration/RoleTable.tsx
@@ -1,10 +1,9 @@
 import { Table, TableBody, TableCell, TableHead, TableRow } from "@mui/material"
 import RoleComponent from "./RoleComponent";
-import RoleComponentXs from "./RoleComponentXS";
 import { RoleTableProps } from "../../models/roleTablePropsInterface";
 
 
-const RoleTable: React.FC<RoleTableProps> = ({ roles, smallSize }) => { // TODO: Corregir la función para recibir un parámetro
+const RoleTable: React.FC<RoleTableProps> = ({ roles }) => { // TODO: Corregir la función para recibir un parámetro
 
   return (
     <Table className="border-table">
@@ -18,11 +17,7 @@ const RoleTable: React.FC<RoleTableProps> = ({ roles, smallSize }) => { // TODO:
       <TableBody>
         {roles && roles.map((role, index) => (
           <TableRow key={index}>
-            {smallSize ? (
               <RoleComponent role={role} />
-            ) : (
-              <RoleComponentXs role={role} />
-            )}
           </TableRow>
         ))}
       </TableBody>
