+++ src/components/common/AddTextModal.tsx
@@ -1,25 +1,20 @@
-import { FC, useEffect, useState } from "react";
-import { Modal as MuiModal, Box, TextField, Button, useMediaQuery, Typography } from "@mui/material";
+import { FC, useState } from "react";
+import { Modal as MuiModal, Box, TextField, Button, Typography } from "@mui/material";
 import { AddTextModalProps } from "../../models/addTextModalPropsInterface";
+
+import './ModalStyle.css';
 
 const AddTextModal: FC<AddTextModalProps> = ({ isVisible, setIsVisible, onCreate }) => {
 const [name, setName] = useState('');
-const isSmallScreen = useMediaQuery('(max-width:500px)');
-
-  useEffect(() => {
-      if (isVisible) {
-        setName('');
-      }
-    }, [isVisible]);
 
   const handleCreate = async () => {
     onCreate(name);
-    setName('');
     setIsVisible(false);
   };
 
   const toggleModal = () => {
     setIsVisible(!isVisible);
+    setName('');
   };
 
   return (
@@ -30,18 +25,7 @@ const isSmallScreen = useMediaQuery('(max-width:500px)');
       aria-describedby="create-modal-description"
     >
       <Box
-        sx={{
-            position: 'absolute',
-            top: '50%',
-            left: '50%',
-            transform: 'translate(-50%, -50%)',
-            width: isSmallScreen ? '90%' : 500,
-            bgcolor: 'background.paper',
-            border: '2px solid #000',
-            boxShadow: 24,
-            p: 4,
-            borderRadius: 1,
-          }}
+        className = 'modal-box'
       >
         <Typography id="create-modal-title" variant = 'h5'>Crear nuevo rol</Typography>
         <TextField
@@ -50,6 +34,7 @@ const isSmallScreen = useMediaQuery('(max-width:500px)');
           onChange={(e) => setName(e.target.value)}
           label="Ingresa el nombre del nuevo rol"
           variant="outlined"
+          inputProps={{ maxLength: 25 }}
           sx={{ marginTop: '20px' }}
         />
         <Box display="flex" justifyContent="flex-end" mt={2}  sx={{ marginTop: '20px' }}>
