+++ src/components/common/ModalStyle.css
@@ -1,0 +1,18 @@
+.modal-box {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 500px;
+  background-color: #fcf9ef;
+  border: 2px solid #000;
+  box-shadow: 24px;
+  padding: 16px;
+  border-radius: 8px;
+}
+
+@media (max-width: 500px) {
+  .modal-box {
+    width: 90%;
+  }
+}
