+++ src/models/roleTablePropsInterface.ts
@@ -2,5 +2,4 @@ import { Role } from "./roleInterface";
 
 export interface RoleTableProps {
     roles: Role[];
-    smallSize: boolean;
     }
